Reduce logging in the message reader (#602)
Motivation:

The message reader gets invoked a lot, especially when messages are
large and split across a number of frames. This generates a lot of
logging. The majority of these logs were `info` and `debug` but in most
cases they're really not that useful.

Modifications:

Change the logging level to `trace` in the message reader.

Result:

Quieter logging.