Generate conformances files containing a message are passed to the plugin (#837)
Motivation:

There was some code in the codegen which checked whether the codegen
request contained any services. If there were no services in the request
then no code would be generated. This was fine when we generated
conformance based on when request/response types of a service. Less see
now.

Modifications:

- Remove the conditional check
- Remove some related dead code
- Add a couple of tests which validate that we do the right thing when
  we generate a file with just a service and the same for just a
  message.

Result:

Less broken codegen