Remove some unnecessary CoWs (#906)
Motivation:

The client state machine holds state in an enum with associated data.
When the associated data is modified it can trigger a CoW for any heap
allocated data since there are two instances of that state: the one held
by the state machine, and the one being modified. This is particularly
bad when reading response data as the relevant states hold a message
reader which in turn holds a buffer accumulating message bytes. Every
time we append to that buffer we end up copying its underlying bytes and
then immediately throwing them away when we update the state held by the
state machine. This is especially bad for large responses which are
delivered in multiple frames.

Modifications:

- Add more benchmarks which avoid networking and return a response in
  multiple frames
- Temporarily switch to a 'modifying' state when modifying the
  underlying state
- When appending to the length prefixed message reader, if the next part
  to read is a message, ensure we reserve enough capacity for the buffer
  to append or the remaining bytes required for the next message to
  read, whichever is larger.

Result:

- Fewer CoWs
- The benchmarks which splits a request over multiple frames saw total
  allocations drop by ~5x.