Tolerate missing content-type header for trailers-only responses (#927)
Motivation:

For trailers-only responses, where the server has 'fast-failed' an RPC,
the spec indicates that a content-type is required and must start with
"application/grpc". However, if this header is missing we will fail the
RPC with a status and message which differ from that sent back from the
server. This is somewhat unhelpful to the end user, as it hides the
real reason for the failure.

Modifications:

- Only validate the content-type for trailers-only responses if it is
  present
- Add tests (and a few additional tests related to trailers-only
  responses)

Result:

- Clients will see the status code and message sent from servers which do
  not include the content-type in their trailers-only responses.
- Resolves #926