Pre-cache the services by name dictionary. (#931)
Motivation:

Currently whenever we construct the GRPCServerRequestRoutingHandler we
need to give it a dictionary of services by name. This is currently done
via a computed property. The result of that is that we compute this
dictionary on the fly every time we receive a new RPC, which is not
exactly the most efficient usage model.

Instead, we should just calculate this ahead of time and save the thing
we actually want, which is the Dictionary. As this data structure is
read vastly more often than it is written, this ends up being a
minor performance win in basically all server applications.

A note: technically this is _very_ subtly breaking as it no longer
guarantees that the Array containing the Services will be in the same
order when you read it as when you set it. I think this is acceptable:
the odds that anyone will rely on this behaviour is low. If we'd minted
1.0.0 I wouldn't take the risk, but right now I think it's ok to take it
and just document our way out of it.

Modifications:

- Replaced stored property storing services with one storing services by
  name.

Results:

Minor performance improvement.