Scaffolding for better server codegen (#1093)
Motivation:

Most of the RPC handling code on the server is generic but not
specialized; as a result we pay a fairly large performance cost. This is
the first in a series of PRs which changes how the shape of the
generated server code.

The plan is to have a new protocol between the generated code and gRPC,
'GRPCServerHandlerProtocol', which operates on concrete types
('ByteBuffer' as opposed to typed request/responses). As such the
protocol for asking for a handler ('CallHandlerProvider') will also
change to return an instance of this new protocol. In generated code the
returned handler will be generic over the request deserializer and
response serializer. The responsibility of the call handler will not
differ much from now (it will receive headers, serialized bytes, etc.)
but will be implemented as a class per call type (this is close to the
current situation where the same is true but a common base class is
shared). In addition the new handlers will not channel handlers: they
will eventually be held by the routing handler.

Modifications:

- Add 'GRPCServerHandlerProtocol', the inbound protocol that the new
  handlers will conform to
- Add 'GRPCServerResponseWriter', the internal outbound counterpart to
  'GRPCServerHandlerProtocol' used by the handlers to communicate back
  to gRPC and add an unimplemented conformance of this to
  'HTTP2ToRawGRPCServerCodec'.
- Add 'UnaryServerHandler' (and tests), a handler for unary requests
- Add 'ServerErrorProcessor'; cargo culted from '_BaseCallHandler'
- Modify the 'CallHandlerProvider' to add a 'handle(method:context:)'

Result:

We have a rough outline of how the new server handlers will look.