Pull channel creation out of `ConnectionManager` (#1158)
Motivation:

The `ConnectionManager` is initialized with a
`ClientConnection.Configuration` struct because it needs enough "stuff"
to make `NIO.Channel`s with. That's not really information it needs to
hold on to, it just needs a way to make channels. Moreover, it limits
the connection manager to using only that configuration object. That's a
bit of a pain if we want to add a connection pool with a different
configuration object.

Modifications:

- Pull out the channel creation into a `ConnectionManagerChannelProvider`
  protocol.
- Move the `ClientConnection.Configuration` based channel creation into
  an object conforming to the above.

Result:

Looser coupling.