Add a static-source Logger (#1165)
Motivation:

swift-log includes a `source` field which -- in most cases -- should be
the name of the module. Currently swift-log parses the `#file` to use
the directory containing the file in the hope that it is the module
name.

In our tests we have a tear down step which validates that all caputred
logs have 'GRPC' as their source. However, this requires manually
providing the source in a bunch of places, that's a bit of a pain and
only validates the logs we emit.

Modifications:

- Add an internal `GRPCLogger` which wraps a `Logger` providing only
  `trace` and `debug` (for now) and always sets the source to `GRPC`.

Result:

It's harder to get the source wrong.