Better handle client sending GOAWAY
Motivation:

The code paths for handling quiescing made an incorrect assumption about
who may open streams while in the quiescing state. This was mostly fine
but fell apart when the client sends a GOAWAY frame to the server and
then opens a stream. This results in the server crashing.

Modifications:

- Remove the unnecessary precondition
- Send a GOAWAY frame to the client if the client sends a GOAWAY frame
  to the server and the server had not previously indicated that it was
  going away. This is followed by a ping whose pong is used as an ack
  that the client has received the GOAWAY and to then ratchet down the
  stream ID in the GOAWAY frame.
- Tests

Result:

Server better handles the client sending it a GOAWAY.