Improve Sendable checking for server code (#1605)
Motivation:

The server handler protocol must be Sendable so that any methods on the
handler are also Sendable as annotating methods as `@Sendable` does not
work as expected. It follows from this that generated server interceptor
factories must also be Sendable (they already are for clients). This
puts the `ServerInterceptor` class in an awkward position: it must be
Sendable but is not inherently thread-safe (these restrictions are
documented and existed before Sendable checking was introduced to
Swift). The `ClientInterceptor` has the same restrictions and is
`@unchecked Sendable` so we elect to do the same for the
`ServerInterceptor`.

Modifications:

- Make generated server handlers and server interceptor factories Sendable
- Make `ServerInterceptor` `@unchecked Sendable`
- Regenerate
- Fix some warnings in test code

Result:

Better Sendable checking