Add SourceFile type for CodeGenLib (#1723)
Motivation:

The CodeRenderer needs to create objects of some type representing the files containing the
generated code. This type is SourceFile.

Modifications:

Implemented the SourceFile struct, containing a name as a String and the generated code as Data.

Result:

The Code Renderer can now be brought from the OpenAPI generator, as it will have an output type.
The code generator will use objects of this type to create the files containing the generated Swift code.