Add a convenience method for creating in-process client/server (#1738)
Motivation:

To use the in-process transport users must create a server and then a
client depending on that server. It rarely makes sense to create one
without the other and I have written more or less the same extension in
three places to do this for me and return a pair. This should just be
made into API.

Modifications:

- Add a `InProcessTransport.makePair()` convenience method
- Replace use of old helpers

Result:

It's easier to create in-process transport.