Use ServiceConfiguration more widely (#1794)
Motivation:

ServiceConfiguration was recently added which has a lot of overlap with
MethodConfigurations. These should be de-duped where possible.

Modifications:

- Make `MethodConfigurations` underscored-public and refactor to use
  `MethodConfiguration.Name`
- Pass `ServiceConfiguration` instead of `MethodConfigurations` where it
  makes sense
- Update docs and tests

* [ProtobufCodeGen] Use the generated input and output type names (#1792)

Motivation:

protoc-gen-swift generates message type names that are different from the ones in the file descriptors. The CodeGenerationRequest object should contain the generated names for the ProtobufCodeGen.

Modifications:

- the parser uses a SwiftProtobufNamer to get the generated type names
and adds them in the CodeGenerationRequest objects
- propagated the change and modified tests accordingly

Result:

The input and output type names will be the generated ones and the generated code will compile.
Co-authored-by: Gustavo Cairo <me@gustavocairo.com>