Fix namespace name for nested package names (#1806)
Motivation:

The current implementation of the ProtobufParser replaces the dots with
their scalar value for the CodeGenerationRequest names. We want the dots
to be replaced by "_"s.

Modifications:

- updated the Name object generation for packages in the ProtobufCodeGenParser to handle correctly nested package names
- added more ProtobufCodeGen tests for a simple package name, a nested one and an empty one

Result:

Tha dots from the package name will be replaced by underscores in the input and output types, as well as in the generated protocols and extensions.