Add call options (#1854)
Motivation:

The API currently doesn't allow for per-method options to be passed to
each call. Users must do so by setting method config on the client. This
isn't particularly ergonomic and also constrains options to the method
config which is insufficient as at a call level users should be able to
control compression as well.

Moreover, these options also need to percolate further down the stack as
they may be used by the transport when creating a stream (compression
and max message size are some examples).

Modifications:

- Add `CallOptions`, which contains the options in method config but
  also includes compression options.
- Percolate these changes down the stack by making them a required
  paramter to the `makeStream` function on the `ClientTransport`.
- Remove the client configuration as the default configuration should be
  set by the transport and overrides are set at the call level.

Result:

- Easier to configure per-call options